reconstruct the codebase structure move all resumetemplates inside components folder
